requirements.txt
addict==2.2.1
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
contextvars==2.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.8
django-appconf==1.0.4
django-braces==1.14.0
django-cors-headers==3.4.0
django-distance-field==1.0.3
django-filter==2.3.0
django-imagekit==4.0
django-leaflet==0.27.0
django-multiselectfield==0.1.12
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-user-agents==0.4.0
djangorestframework==3.10.0
djangorestframework-gis==0.15
drf-autodocs @ git+https://github.com/michameiu/drf-autodocs.git@556e8c562190797eed26c284c4a0400bed1bab0e
et-xmlfile==1.0.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
importlib-metadata==1.7.0
jdcal==1.4.1
Markdown==3.2.2
oauthlib==3.1.0
onesignal-sdk==2.0.0
openpyxl==3.0.4
pilkit==2.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
zipp==3.1.0
